chore: build appx in another ci step (#77)
chore: build appx

+++ scripts/prepare-artifacts.js
@@ -2,13 +2,18 @@ const fs = require('fs');
 const dir = './output';
 const ext = process.argv[2];
 console.log('ext: ', ext);
-if (!fs.existsSync(dir)){
+if (!fs.existsSync(dir)) {
     fs.mkdirSync(dir);
 }
 
 if (ext === 'exe') {
-    fs.copyFileSync(`dist/Install Decentraland.${ext}`, `output/Install Decentraland.${ext}`)
-    fs.copyFileSync(`dist/Decentraland.appx`, `output/Decentraland.appx`)
+    if (fs.existsSync(`dist/Install Decentraland.${ext}`)) {
+        fs.copyFileSync(`dist/Install Decentraland.${ext}`, `output/Install Decentraland.${ext}`)
+    }
+
+    if (fs.existsSync(`dist/Decentraland.appx`)) {
+        fs.copyFileSync(`dist/Decentraland.appx`, `output/Decentraland.appx`)
+    }
 } else {
     fs.copyFileSync(`dist/Decentraland.${ext}`, `output/Decentraland.${ext}`)
 }
